+++ ChangeLog.md
@@ -1,6 +1,8 @@
 0.3.0 [2020-08-06]
 
  * Add --print-groups option to represent
+ * Only consider dates if --max-day-sep argument is given (no default, but still sort by date)
+ * Do not parse states unless --same-state is given 
 
 0.2.0 [2020-01-06]